feat:backend routes updated

+++ backend/src/routes/exams.ts
@@ -10,7 +10,7 @@ router.get('/', authenticateJWT, collegeIsolation, async (req: AuthRequest, res:
     try {
         let sql = `
             SELECT e.*, 
-                   (SELECT COUNT(*) FROM questions WHERE exam_id = e.id) as question_count,
+                   (SELECT COUNT(*)::int FROM questions WHERE exam_id = e.id) as question_count,
                    EXISTS (SELECT 1 FROM attempts WHERE exam_id = e.id AND student_id = $1 AND submitted_at IS NOT NULL) as is_attempted,
                    (SELECT score FROM attempts WHERE exam_id = e.id AND student_id = $1 ORDER BY created_at DESC LIMIT 1) as score
             FROM exams e
